% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dv.R
\name{add_dv}
\alias{add_dv}
\title{Add a dependent variable to a fixed-effect model}
\usage{
add_dv(
  .data,
  vars = c(dv = "Dependent Variable"),
  between = list(),
  mean = 0,
  sd = 1,
  r = 0,
  dist = "norm",
  params = list(),
  empirical = FALSE
)
}
\arguments{
\item{.data}{the data frame}

\item{vars}{a vector of the DV(s) to create}

\item{between}{specifies the between cells}

\item{mean}{the mean(s) for the DV(s) - ignored if params is set}

\item{sd}{the SD(s) for the DV(s) - ignored if params is set}

\item{r}{the correlation for the DVs (if >1 DV)}

\item{dist}{the distribution(s) for the DVs}

\item{params}{the distribution parameters}

\item{empirical}{logical. If true, mu, sd, r and params specify the empirical not population statistics}
}
\value{
a data frame with new DV column(s)
}
\description{
Add a dependent variable to a fixed-effect model
}
\examples{
df <- add_random(id = 1000) |>
  add_between(condition = c("ctl", "exp")) |>
  add_dv(vars = c(pre = "Pre-test Score", post = "Post-test Score"),
         between = list(condition = "ctl"),
         mean = c(100, 105),
         sd = 10,
         r = 0.5,
         ) |>
  add_dv(vars = c(pre = "Pre-test Score", post = "Post-test Score"),
         between = list(condition = "exp"),
         mean = c(100, 110),
         sd = c(10, 8),
         r = 0.4)
}
