% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add column labels}
\usage{
add_labels(.data, ...)
}
\arguments{
\item{.data}{the data frame}

\item{...}{the column names and lables (e.g., \code{id = "Student ID"})}
}
\value{
data frame with column labels
}
\description{
Add column labels
}
\examples{
df <- add_random(rid = 2, sid = 2) |>
  add_labels(rid = "Rater ID", sid = "Stimulus ID")
View(df)
}
