% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{add_labels_from_design}
\alias{add_labels_from_design}
\title{Add Labels to Data Tables}
\usage{
add_labels_from_design(data, design = NULL)
}
\arguments{
\item{data}{a data table}

\item{design}{a design specification (if NULL, is read from data)}
}
\value{
a data frame with labelled columns
}
\description{
Add Labels to Data Tables
}
\keyword{internal}
