% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ranef.R
\name{add_within}
\alias{add_within}
\title{Add within factors}
\usage{
add_within(.data, .by = NULL, ...)
}
\arguments{
\item{.data}{the data frame}

\item{.by}{the grouping column (groups by row if NULL)}

\item{...}{the names and levels of the new factors}
}
\value{
data frame
}
\description{
Add within factors
}
\examples{
add_random(subj = 2, item =  2) \%>\%
  add_within("subj", time = c("pre", "post"))
}
