% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_convertors.R
\name{average_r2tau_0}
\alias{average_r2tau_0}
\title{Average r to Random Intercept SD}
\usage{
average_r2tau_0(average_r, sigma)
}
\arguments{
\item{average_r}{The average inter-item correlation}

\item{sigma}{Total error variance}
}
\value{
The standard deviation of the random intercept
}
\description{
Average r to Random Intercept SD
}
