% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{cell_combos}
\alias{cell_combos}
\title{Cell combos}
\usage{
cell_combos(factors)
}
\arguments{
\item{factors}{A list of factors}
}
\value{
a list
}
\description{
Creates wide cell combination names, such as A1_B1, A2_B1, A1_B2, A2_B2.
}
\keyword{internal}
