% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mixed_design.R
\name{check_mixed_design}
\alias{check_mixed_design}
\title{Get random intercepts for subjects and items}
\usage{
check_mixed_design(data, dv = 1, sub_id = 2, item_id = 3, formula = NULL)
}
\arguments{
\item{data}{the existing tbl}

\item{dv}{the column name or index containing the DV}

\item{sub_id}{the column name or index for the subject IDs}

\item{item_id}{the column name or index for the item IDs}

\item{formula}{the formula to run in lmer (defaults to null model dv ~ 1 + (1|sub_id) + (1|item_id))}
}
\value{
a list of parameters
}
\description{
Get error terms from an existing data table.
}
\examples{
des <- check_mixed_design(fr4, "rating", "rater_id", "face_id")
str(des[1:4])
}
