% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_stats.R
\name{check_sim_stats}
\alias{check_sim_stats}
\title{Check table stats}
\usage{
check_sim_stats(dat, within = c(), between = c(), dv = c(),
  id = c(), digits = 2, usekable = FALSE, grp_by = between)
}
\arguments{
\item{dat}{the existing dataframe}

\item{digits}{how many digits to round to (default = 2)}

\item{usekable}{logical. If TRUE, output with knitr::kable}

\item{grp_by}{an optional list of column names to group by}
}
\value{
tibble or kable
}
\description{
\code{check_sim_stats} Generates a table of the correlations and means of numeric columns in a data frame
}
\examples{
check_sim_stats(iris, "Species")
}
