% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkstats.R
\name{checkstats}
\alias{checkstats}
\title{Check table stats}
\usage{
checkstats(dat, grp_by = NULL, digits = 2, usekable = FALSE)
}
\arguments{
\item{dat}{the existing dataframe}

\item{grp_by}{an optional list of column names to group by}

\item{digits}{how many digits to round to (default = 2)}

\item{usekable}{logical. If TRUE, output with knitr::kable}
}
\value{
tibble or kable
}
\description{
\code{checkstats} Generates a table of the correlations and means of numeric columns in a data frame
}
\examples{
checkstats(iris, "Species")
}
