% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_interactive.R
\name{codebook_interactive}
\alias{codebook_interactive}
\title{Interactive Codebook}
\usage{
codebook_interactive(data, cb = NULL)
}
\arguments{
\item{data}{The data frame to generate a codebook for}

\item{cb}{The codebook in list format if already generated}
}
\value{
codebook list
}
\description{
Create a Psych-DS formatted codebook from data by answering questions interactively in the console.
}
\keyword{internal}
