% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrast_code}
\alias{contrast_code}
\title{Contrast code a factor}
\usage{
contrast_code(fct, levels = NULL, base = 1)
}
\arguments{
\item{fct}{the factor to contrast code (or a vector)}

\item{levels}{the levels of the factor in order}

\item{base}{the index of the level to use as baseline}
}
\value{
the factor with contrasts set
}
\description{
Contrast coding sets the grand mean as the intercept. 
Each contrast compares one level with the reference level (base).
}
\examples{
df <- sim_design(between = list(pet = c("cat", "dog")), 
                 mu = c(10, 20), plot = FALSE)
df$pet <- contrast_code(df$pet)
lm(y ~ pet, df) \%>\% broom::tidy()

df <- sim_design(between = list(pet = c("cat", "dog", "ferret")), 
                 mu = c(2, 4, 9), empirical = TRUE, plot = FALSE)
                 
df$pet <- contrast_code(df$pet, base = 1)
lm(y ~ pet, df) \%>\% broom::tidy()

df$pet <- contrast_code(df$pet, base = 2)
lm(y ~ pet, df) \%>\% broom::tidy()

df$pet <- contrast_code(df$pet, base = "ferret")
lm(y ~ pet, df) \%>\% broom::tidy()
}
