% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{convert_param}
\alias{convert_param}
\title{Convert mu or sd}
\usage{
convert_param(param, cells_b, cells_w, type = "this parameter")
}
\arguments{
\item{param}{the parameter (mu, sd, or n)}

\item{cells_b}{a list of between-subject cell combinations}

\item{cells_w}{a list of within-subject cells combinations}

\item{type}{the name of the parameter (for error messages)}
}
\value{
a data frame
}
\description{
Converts parameter specification from vector or list format
}
