% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Make a correlation matrix}
\usage{
cormat(cors = 0, vars = 3)
}
\arguments{
\item{cors}{the correlations among the variables (can be a single number,
vars$\times$vars matrix, vars$\times$vars vector, or a vars$\times$(vars-1)/2 vector)}

\item{vars}{the number of variables in the matrix}
}
\value{
matrix
}
\description{
\code{cormat} makes a correlation matrix from a single number,
vars$\times$vars matrix, vars$\times$vars vector, or a vars$\times$(vars-1)/2 vector.
}
\examples{
cormat(.5, 3)
cormat(c( 1, .2, .3, .4,
         .2,  1, .5, .6, 
         .3, .5,  1, .7,
         .4, .6, .7,  1), 4)
cormat(c(.2, .3, .4, .5, .6, .7), 4)
}
