% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat_from_triangle}
\alias{cormat_from_triangle}
\title{Make Correlation Matrix from Triangle}
\usage{
cormat_from_triangle(cors)
}
\arguments{
\item{cors}{the correlations among the variables as a vars\*(vars-1)/2 vector}
}
\value{
matrix
}
\description{
\code{cormat_from_triangle} makes a correlation matrix from a vector of the upper right triangle
}
\examples{
cormat_from_triangle(c(.2, .3, .4, 
                           .5, .6, 
                               .7))
}
