% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{difference_code}
\alias{difference_code}
\title{Difference code a factor}
\usage{
difference_code(fct, levels = NULL)
}
\arguments{
\item{fct}{the factor to contrast code (or a vector)}

\item{levels}{the levels of the factor in order}
}
\value{
the factor with contrasts set
}
\description{
Difference coding sets the grand mean as the intercept. 
Each contrast compares one level with the previous level.
}
\examples{
df <- sim_design(between = list(pet = c("cat", "dog", "ferret")), 
                 mu = c(2, 4, 9), empirical = TRUE, plot = FALSE)
                 
df$pet <- difference_code(df$pet)
lm(y ~ pet, df) \%>\% broom::tidy()

}
