% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_options.R
\name{faux_options}
\alias{faux_options}
\title{Set/get global faux options}
\usage{
faux_options(...)
}
\arguments{
\item{...}{One of four: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing), (4) a list with option-value pairs which sets all the corresponding arguments.}
}
\value{
a list of options, values of an option, or nothing
}
\description{
Global faux options are used, for example, to set the default separator for cell names.
}
\examples{

faux_options() # see all options

faux_options("sep") # see value of faux.sep

\donttest{
# changes cell separator (e.g., A1.B2)
faux_options(sep = ".")

# changes cell separator back to default (e.g., A1_B2)
faux_options(sep = "_")
}
}
