% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faceratings.R
\docType{data}
\name{fr10}
\alias{fr10}
\title{Attractiveness ratings of 10 faces from 10 subjects}
\format{A data frame with 100 rows and 9 variables:
\describe{
  \item{rater_id}{rater's ID}
  \item{rater_sex}{rater's sex (female, male, intersex, NA)}
  \item{rater_age}{rater's age (17-90 years)}
  \item{rater_sexpref}{rater's preferred sex for romantic relationships (either, men, neither, women, NA)}
  \item{face_id}{face's ID}
  \item{face_sex}{face's sex (female, male)}
  \item{face_age}{face's age (in years)}
  \item{face_eth}{face's ethnic group}
  \item{rating}{attractiveness rating on a scale from 1 (much less attractive than average) to 7 (much more attractive than average)}
}}
\source{
\url{https://figshare.com/articles/Face_Research_Lab_London_Set/5047666}
}
\usage{
fr10
}
\description{
The faceratings dataset cut down to the first 10 raters and the first 10 faces for demos
}
\keyword{datasets}
