% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{get_contrast_vals}
\alias{get_contrast_vals}
\title{Get contrast values}
\usage{
get_contrast_vals(v)
}
\arguments{
\item{v}{a factor vector}
}
\value{
a data frame
}
\description{
Get a data frame of contrast values from a factor vector
}
\examples{
dat <- sim_design(
  between = list(group = c("A", "B")), 
  n = 5, plot = FALSE)

get_contrast_vals(dat$group)

}
