% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{get_mu_sd}
\alias{get_mu_sd}
\title{Convert mu or sd}
\usage{
get_mu_sd(param, cells_b, cells_w, type = "this parameter")
}
\arguments{
\item{param}{the mu or sd}

\item{cells_b}{a list of between-subject cell combinations}

\item{cells_w}{a list of within-subject cells combinations}

\item{type}{the name of the parameter (for error messages)}
}
\value{
a data frame
}
\description{
Checks mu and sd specification from vector or list format
}
