% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcols.R
\name{getcols}
\alias{getcols}
\title{Get data columns}
\usage{
getcols(data, ..., as_index = FALSE)
}
\arguments{
\item{data}{}

\item{...}{}

\item{as}{}
}
\value{

}
\description{
Get columns from a data table by specifying the index, column name as a string, or unquoted column name. Returns the column names or indices.
}
\examples{
getcols(mtcars, 1L, cyl, "disp", 4.0)
}
