% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{interactive_codebook}
\alias{interactive_codebook}
\title{Interactive Codebook}
\usage{
interactive_codebook(data, cb = NULL)
}
\arguments{
\item{data}{The data frame to generate a codebook for}

\item{cb}{The codebook in list format if already generated}
}
\value{

}
\description{
Interactive Codebook
}
