% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{is_pos_def}
\alias{is_pos_def}
\title{Check a Matrix is Positive Definite}
\usage{
is_pos_def(cor_mat, tol = 1e-08)
}
\arguments{
\item{cor_mat}{a correlation matrix}

\item{tol}{the tolerance for comparing eigenvalues to 0}
}
\value{
logical value
}
\description{
\code{is_pos_def} makes a correlation matrix from a vector
}
\examples{
is_pos_def(matrix(c(1, .5, .5, 1), 2)) # returns TRUE
is_pos_def(matrix(c(1, .9, .9, 
                   .9, 1, -.2, 
                   .9, -.2, 1), 3)) # returns FALSE
}
