% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message}
\alias{message}
\title{Less scary green messages}
\usage{
message(..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{...}{message components (see \code{\link[base:message]{base::message()}})}

\item{domain}{(see \code{\link[base:message]{base::message()}})}

\item{appendLF}{append new line? (see \code{\link[base:message]{base::message()}})}
}
\value{
TRUE
}
\description{
Less scary green messages
}
\keyword{internal}
