% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy.R
\name{messy}
\alias{messy}
\title{Simulate missing data}
\usage{
messy(data, prop = 0, ..., replace = NA)
}
\arguments{
\item{data}{the tbl}

\item{prop}{the proportion of data to mess up}

\item{...}{the columns to mess up (as a vector of column names or numbers)}

\item{replace}{the replacement value (defaults to NA)}
}
\value{
the messed up table
}
\description{
Insert NA or another replacement value for some proportion of specified
columns to simulate missing data.
}
\examples{
messy(iris, 0.1, "Species", replace = "NO SPECIES")
messy(iris, 0.5, 1:4)
}
