% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnormpre.R
\name{mvrnormpre}
\alias{mvrnormpre}
\title{Simulate from a Multivariate Normal Distribution with One Pre-Existing Vector}
\usage{
mvrnormpre(pre = NULL, n = 1, mu, Sigma, tol = 1e-06,
  empirical = FALSE)
}
\arguments{
\item{n}{the number of smaples required}

\item{mu}{a vector giving the means of the variables}

\item{Sigma}{a positive-definite symmetric matrix specifying the covariance matrix of the variables.}

\item{empirical}{logical. If true, mu and Sigma specify the empirical not population mean and covariance}
}
\value{
matrix
}
\description{
\code{multirnorm} Produces one or more samples from the specified multivariate normal distribution, based on one pre-existing vector
}
