% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{nested_list}
\alias{nested_list}
\title{Output a nested list in RMarkdown list format}
\usage{
nested_list(x, pre = "")
}
\arguments{
\item{x}{The list}

\item{pre}{Test to prefix to each line (e.g., if you want all lines indented 4 spaces to start, use "    ")}
}
\value{
A character string
}
\description{
Output a nested list in RMarkdown list format
}
\examples{
x <- list(
  a = list(a1 = "Named", a2 = "List"),
  b = list("Unnamed", "List"),
  c = c(c1 = "Named", c2 = "Vector"),
  d = c("Unnamed", "Vector"),
  e = list(e1 = list("A", "B", "C"),
           e2 = list(a = "A", b = "B"),
           e3 = c("A", "B", "C"),
           e4 = 100),
  f = "not a list or vector"
)
nested_list(x)
}
