% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of applying the function in the rhs to the value in the lhs.
}
\description{
See \verb{dplyr::[\\\%>\\\%][dplyr::\\\%>\\\%]} for details.
}
\keyword{internal}
