% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{poly_code}
\alias{poly_code}
\title{Polynomial code a factor}
\usage{
poly_code(fct, levels = NULL)
}
\arguments{
\item{fct}{the factor to contrast code (or a vector)}

\item{levels}{the levels of the factor in order}
}
\value{
the factor with contrasts set
}
\description{
Polynomial code a factor
}
\examples{
df <- sim_design(within = list(time = 1:6), 
                 mu = 1:6 + (1:6-3.5)^2, 
                 long = TRUE, plot = FALSE)
                 
df$time <- poly_code(df$time)
lm(y ~ time, df) \%>\% broom::tidy()

}
