% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{pos_def_limits}
\alias{pos_def_limits}
\title{Limits on Missing Value for Positive Definite Matrix}
\usage{
pos_def_limits(..., steps = 0.01, tol = 1e-08)
}
\arguments{
\item{...}{the correlations among the variables as a vars\*(vars-1)/2 vector}

\item{steps}{the tolerance for min and max values}

\item{tol}{the tolerance for comparing eigenvalues to 0}
}
\value{
dataframe with min and max values
}
\description{
\code{pos_def_limits} returns min and max possible values for a positive definite matrix with a specified missing value
}
\examples{
pos_def_limits(.8, .2, NA)
}
