% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{print.design}
\alias{print.design}
\title{Print Design List}
\usage{
\method{print}{design}(x, ...)
}
\arguments{
\item{x}{The design list}

\item{...}{Additional parameters for print}
}
\description{
Print Design List
}
