% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{print.nested_list}
\alias{print.nested_list}
\title{Print Codebook Object}
\usage{
\method{print}{nested_list}(x, ...)
}
\arguments{
\item{x}{The nested_list string}

\item{...}{Additional parameters for print}
}
\description{
Print Codebook Object
}
