% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{print.psychds_codebook}
\alias{print.psychds_codebook}
\title{Print Codebook Object}
\usage{
\method{print}{psychds_codebook}(x, ...)
}
\arguments{
\item{x}{The psychds_codebook list}

\item{...}{Additional parameters for print}
}
\description{
Print Codebook Object
}
\keyword{internal}
