% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormpre.R
\name{rnormpre}
\alias{rnormpre}
\title{Correlated Normal Vector}
\usage{
rnormpre(x, rho = 0, ymean = 0, ysd = 1)
}
\arguments{
\item{x}{the existing vector}

\item{rho}{desired correlation between existing and returned vectors}

\item{ymean}{desired mean of returned vector}

\item{ysd}{desired SD of returned vector}
}
\value{
vector
}
\description{
\code{rnormpre} Produces a random normally distributed vector with the specified correlation to an existing vector
}
\examples{
v1 <- rnorm(10)
v2 <- rnormpre(v1, 0.5, 0, 1)
cor(v1, v2)
}
