% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_pre.R
\name{sample_from_pop}
\alias{sample_from_pop}
\title{Sample Parameters from Population Parameters}
\usage{
sample_from_pop(n = 100, mu = 0, sd = 1, r = 0)
}
\arguments{
\item{n}{sample size}

\item{mu}{population mean}

\item{sd}{population SD}

\item{r}{population r}
}
\value{
list of sample parameters (mu, sd, r)
}
\description{
Sample Parameters from Population Parameters
}
\examples{
sample_from_pop(10)
}
