% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_num_grp.R
\name{select_num_grp}
\alias{select_num_grp}
\title{Select grouping and numeric columns and group}
\usage{
select_num_grp(dat, grp_by = NULL)
}
\arguments{
\item{dat}{the existing dataframe}

\item{grp_by}{an optional list of column names to group by}
}
\value{
tibble
}
\description{
\code{select_num_grp} Select grouping and numeric columns and group
}
\examples{
select_num_grp(iris, "Species")
}
