% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{set_design}
\alias{set_design}
\title{Set design}
\usage{
set_design(data, design)
}
\arguments{
\item{data}{The data table}

\item{design}{The design list}
}
\value{
A data frame with a design attribute
}
\description{
Add a design specification to a data table
}
\examples{
design <- check_design()
data <- data.frame(id = 1:100, y = rnorm(100)) \%>\%
  set_design(design)
}
