% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sim_seed.R
\name{set_sim_seed}
\alias{set_sim_seed}
\title{Set seed for a simulated data set}
\usage{
set_sim_seed(seed, kind = "Mersenne-Twister",
  normal.kind = "Inversion")
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or NULL (see set.seed)}

\item{kind}{(see set.seed)}

\item{normal.kind}{(see set.seed)}
}
\value{
sets the seed
}
\description{
Set seed for a simulated data set
}
\keyword{internal}
