% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from design (internal)}
\usage{
sim_data(design, empirical = FALSE, long = FALSE, seed = NULL,
  sep = faux_options("sep"))
}
\arguments{
\item{design}{A list of design parameters created by check_design()}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}

\item{seed}{a single value, interpreted as an integer, or NULL (see set.seed)}

\item{sep}{separator for within-columns, defaults to _ (see tidyr::separate)}
}
\value{
a tbl
}
\description{
Simulate data from design (internal)
}
