% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design_}
\alias{sim_design_}
\title{Simulate Data from Design}
\usage{
sim_design_(design, empirical = FALSE, long = FALSE)
}
\arguments{
\item{design}{A list of design parameters created by check_design()}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}
}
\value{
a tbl
}
\description{
\code{sim_design_} generates a data table with a specified design
}
\keyword{internal}
