% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_joint_dist.R
\name{sim_joint_dist}
\alias{sim_joint_dist}
\title{Simulate category joint distribution}
\usage{
sim_joint_dist(data, ..., n = 100, empirical = FALSE)
}
\arguments{
\item{data}{the existing tbl}

\item{...}{columns to calculate the joint distribution from, if none are chosen, all columns with 10 or fewer unique values will be chosen}

\item{n}{the number of total observations to return}

\item{empirical}{Should the returned data have the exact same distribution of conditions? (versus be sampled from a population with this distribution)}
}
\value{
data table
}
\description{
This function is mainly used internally, such as for simulating  missing data patterns, but is available in case anyone finds it useful.
}
\examples{
sim_joint_dist(ggplot2::diamonds, cut, color, n = 10)
}
