% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{sim_params}
\alias{sim_params}
\title{Get simulation parameters from a data table}
\usage{
sim_params(data)
}
\arguments{
\item{data}{the existing tbl}
}
\value{
a tbl of correlations, means and sds
}
\description{
Generates a table of the correlations and means of numeric columns in a data frame, if data was generated by \link{sim_design} and has a \code{design} attribute
}
\examples{
dat <- sim_design(c(2, 2), c(2, 2), mu = 1:16, sd = 2)
sim_params(dat)
sample_params(dat)
}
