% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdf.R
\name{simdf}
\alias{simdf}
\title{Simulate an existing dataframe}
\usage{
simdf(dat, n = 100, grp_by = NULL)
}
\arguments{
\item{dat}{the existing dataframe}

\item{n}{the number of samples to return per group}

\item{grp_by}{an optional list of column names to group by}
}
\value{
tibble
}
\description{
\code{simdf} Produces a dataframe with the same distributions and correlations as an existing dataframe. Only returns numeric columns and simulates all numeric variables from a continuous normal distribution (for now).
}
\examples{
iris100 <- simdf(iris, 100)
iris_species <- simdf(iris, 100, "Species")
}
