% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdf_mixed.R
\name{simdf_mixed}
\alias{simdf_mixed}
\title{Generate a sample with random intercepts for subjects and items}
\usage{
simdf_mixed(dat, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)
}
\arguments{
\item{dat}{the existing dataframe}

\item{sub_n}{the number of subjects to simulate}

\item{item_n}{the number of items to simulate}

\item{dv}{the column name or index containing the DV}

\item{sub_id}{the column name or index for the subject IDs}

\item{item_id}{the column name or index for the item IDs}
}
\value{
tibble
}
\description{
\code{simdf_mixed} Produces a dataframe with the same distributions of by-subject and by-item random intercepts as an existing dataframe
}
\examples{
\donttest{simdf_mixed(faceratings, 10, 10, "rating", "rater_id", "face_id")}
}
