% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_convertors.R
\name{std_alpha2average_r}
\alias{std_alpha2average_r}
\title{Standardized Alpha to Average R}
\usage{
std_alpha2average_r(std_alpha, n)
}
\arguments{
\item{std_alpha}{The standarized alpha}

\item{n}{The number of items}
}
\value{
The average inter-item correlation
}
\description{
Standardized Alpha to Average R
}
\examples{
std_alpha2average_r(.8, 10)
}
