% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{sum_code}
\alias{sum_code}
\title{Sum code a factor}
\usage{
sum_code(fct, levels = NULL, omit = length(levels))
}
\arguments{
\item{fct}{the factor to contrast code (or a vector)}

\item{levels}{the levels of the factor in order}

\item{omit}{the level to omit (defaults to the last level)}
}
\value{
the factor with contrasts set


}
\description{
Sum coding sets the grand mean as the intercept. 
Each contrast compares one level with the grand mean.
}
\examples{
df <- sim_design(between = list(pet = c("cat", "dog", "ferret")), 
                 mu = c(2, 4, 9), empirical = TRUE, plot = FALSE)

df$pet <- sum_code(df$pet)
lm(y ~ pet, df) \%>\% broom::tidy()

df$pet <- sum_code(df$pet, omit = "cat")
lm(y ~ pet, df) \%>\% broom::tidy()

df$pet <- sum_code(df$pet, omit = 1)
lm(y ~ pet, df) \%>\% broom::tidy()
}
