% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{unique_pairs}
\alias{unique_pairs}
\title{Make unique pairs of level names for correlations}
\usage{
unique_pairs(v)
}
\arguments{
\item{v}{a vector of level names or a number of levels}
}
\value{
a vector of all unique pairs
}
\description{
Make unique pairs of level names for correlations
}
\examples{
unique_pairs(c("O", "C", "E", "A", "N"))
unique_pairs(3)
}
