% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{wide2long}
\alias{wide2long}
\title{Wide to long format}
\usage{
wide2long(dat, within_factors = c(), within_cols = c(),
  sep = "[^[:alnum:]]+")
}
\arguments{
\item{dat}{the wide data frame to convert}

\item{sep}{Separator between columns (see tidyr::separate)}

\item{within}{the names of the within factors}

\item{dv}{the names of the DV (value) columns}
}
\value{
the data frame in long format
}
\description{
Converts data from wide format to long
}
