% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dv.R
\name{add_dv}
\alias{add_dv}
\title{Add a dependent variable}
\usage{
add_dv(
  .data,
  formula = y ~ 1,
  intercept = 0,
  error = 1,
  fixef = list(),
  ranef = list()
)
}
\arguments{
\item{.data}{the data frame}

\item{formula}{The formula for your model}

\item{intercept}{The (grand) intercept value}

\item{error}{The SD of the error term}

\item{fixef}{A list of fixed effects (see Details)}

\item{ranef}{A list of random effects parameters (see Details)}
}
\value{
a data frame with new DV column
}
\description{
Add a dependent variable to a mixed model simulation using a formula and specification of fixed and random effects parameters.
}
\details{
Fixed effects are specified as a named list for each effect in the equation. For example, for the equation `y ~ a * b + (1 | id)`, the fixed effects might be specified as such: `list(a = 5, b = 10, "a:b" = 0)`.

Random effects are also specified as a named list of standard deviations for the random intercept and slopes, plus optional correlations. For example, for the equation `y ~ a * b + (b | id)`, the random effects might be specified as such: `list(id = list(intercept = 10, b = 5, .cors = 0.4))`.
}
\examples{
add_random(id = 1000) |> 
  add_between(a = c("A1", "A2")) |>
  add_within(b = c("B1", "B2")) |>
  add_dv(y ~ a*b + (b | id),
         intercept = 100,
         error = 10,
         fixef = list(a = 5, b = 10, "a:b" = 0),
         ranef = list(id = list(intercept = 10, b = 5, .cors = 0.4))
  )
}
