% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_combos.R
\name{cell_combos}
\alias{cell_combos}
\title{Cell combos}
\usage{
cell_combos(factors, dv = "y", sep = faux_options("sep"))
}
\arguments{
\item{factors}{A list of lists of named factor levels}

\item{dv}{name of dv column ("y") to be used if there are no factors}

\item{sep}{separator}
}
\value{
a list
}
\description{
Creates wide cell combination names, such as A1_B1, A2_B1, A1_B2, A2_B2.
}
\keyword{internal}
