% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{check_design}
\alias{check_design}
\title{Validate design}
\usage{
check_design(within = list(), between = list(), n = 100, cors = 0,
  mu = 0, sd = 1, tmpvar_name = ".tmpvar.")
}
\arguments{
\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{cors}{the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)}

\item{mu}{a vector giving the means of the variables (numeric vector of length 1 or vars)}

\item{sd}{the standard deviations of the variables (numeric vector of length 1 or vars)}

\item{tempvar_name}{the name to assign placeholders when there are no within or no between factors}
}
\value{
list
}
\description{
\code{check_design} validates the specified within and between design
}
