% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{check_design}
\alias{check_design}
\title{Validate design}
\usage{
check_design(within = list(), between = list(), n = 100, mu = 0,
  sd = 1, r = 0, plot = TRUE)
}
\arguments{
\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{mu}{a vector giving the means of the variables (numeric vector of length 1 or vars)}

\item{sd}{the standard deviations of the variables (numeric vector of length 1 or vars)}

\item{r}{the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)}

\item{plot}{whether to show a plot of the design}
}
\value{
list
}
\description{
\code{check_design} validates the specified within and between design
}
\examples{

within <- list(time = c("day", "night"))
between <- list(pet = c("dog", "cat"))
check_design(within, between)

}
