% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_stats.R
\name{check_sim_stats}
\alias{check_sim_stats}
\title{Check table stats}
\usage{
check_sim_stats(dat, between = c(), within = c(), dv = c(),
  id = c(), digits = 2, usekable = FALSE)
}
\arguments{
\item{dat}{the existing dataframe}

\item{between}{a vector of column names for between-subject factors}

\item{within}{a vector of column names for within-subject factors (if data is long)}

\item{dv}{the column name of the dv (if data is long)}

\item{id}{the column name(s) of the subject ID (if data is long)}

\item{digits}{how many digits to round to (default = 2)}

\item{usekable}{logical. If TRUE, output with knitr::kable}
}
\value{
tibble or kable
}
\description{
\code{check_sim_stats} Generates a table of the correlations and means of numeric columns in a data frame
}
\examples{
check_sim_stats(iris, "Species")
}
