% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_stats.R
\name{check_sim_stats}
\alias{check_sim_stats}
\title{Check table stats}
\usage{
check_sim_stats(dat, grp_by = NULL, digits = 2, usekable = FALSE)
}
\arguments{
\item{dat}{the existing dataframe}

\item{grp_by}{an optional list of column names to group by}

\item{digits}{how many digits to round to (default = 2)}

\item{usekable}{logical. If TRUE, output with knitr::kable}
}
\value{
tibble or kable
}
\description{
\code{check_sim_stats} Generates a table of the correlations and means of numeric columns in a data frame
}
\examples{
check_sim_stats(iris, "Species")
}
