% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create PsychDS Codebook from Data}
\usage{
codebook(data, coldesc = NULL, as_json = TRUE, interactive = FALSE)
}
\arguments{
\item{data}{The data frame to generate a codebook for}

\item{coldesc}{Optional named list of column descriptions}

\item{as_json}{Whether the output should be a list or JSON format}

\item{interactive}{Whether the function should prompt the user to describe columns and factor levels}
}
\value{
a list or json-formatted codebook
}
\description{
Create PsychDS Codebook from Data
}
\examples{

data <- sim_design(2, 2, long = TRUE)
cb <- codebook(data, c("A" = "Factor 1", "B" = "Factor 2"))
}
