% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_param}
\alias{convert_param}
\title{Convert parameter}
\usage{
convert_param(param, cells_w, cells_b, type = "this parameter")
}
\arguments{
\item{param}{the parameter (mu, sd, or n)}

\item{cells_w}{a vector of within-subject cells combinations}

\item{cells_b}{a vector of between-subject cell combinations}

\item{type}{the name of the parameter (for error messages)}
}
\value{
a list of parameters
}
\description{
Converts parameter specification from vector or list format
}
\keyword{internal}
