% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux-package.R
\docType{package}
\name{faux}
\alias{faux}
\alias{faux-package}
\title{faux: Simulation Functions.}
\description{
The faux package provides functions for simulating datasets with specified structure.
}
\keyword{package}
