% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{fix_name_labels}
\alias{fix_name_labels}
\title{Fix name labels}
\usage{
fix_name_labels(x, replace = "(\\\\W|_)")
}
\arguments{
\item{x}{the list to fix}

\item{replace}{regex pattern to replace with full stops (defaults to non-word characters and underscores)}
}
\value{
the fixed list
}
\description{
Fixes if a factor list does not have named levels or has special characters in the names
}
\keyword{internal}
