% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{fix_name_labels}
\alias{fix_name_labels}
\title{Fix name labels}
\usage{
fix_name_labels(x)
}
\arguments{
\item{x}{the list to fix}
}
\value{
the fixed list
}
\description{
Fixes if a factor list does not have named levels or has special characters in the names
}
\keyword{internal}
