% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{get_design_long}
\alias{get_design_long}
\title{Get design from long data}
\usage{
get_design_long(.data, id = "sub_id", dv = "val")
}
\arguments{
\item{.data}{the data frame (in long format)}

\item{id}{the column name(s) that identify a unit of analysis}

\item{dv}{the column name that identifies the DV}
}
\value{
the data frame in long format
}
\description{
Makes a best guess at the design of a long-format data frame. 
Finds all columns that contain a single value per unit of analysis (between factors), 
all columns that contain the same values per unit of analysis (within factors), and 
all columns that differ over units of analysis (dv, continuous factors)
}
