% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\alias{check_sim_stats}
\title{Get parameters from a data table}
\usage{
get_params(
  data,
  between = NULL,
  within = NULL,
  dv = NULL,
  id = NULL,
  digits = 2
)

check_sim_stats(
  data,
  between = NULL,
  within = NULL,
  dv = NULL,
  id = NULL,
  digits = 2
)
}
\arguments{
\item{data}{the existing tbl}

\item{between}{a vector of column names for between-subject factors}

\item{within}{a vector of column names for within-subject factors (if data is long)}

\item{dv}{the column name(s) of the dv, if NULL all numeric columns will be selected}

\item{id}{the column name(s) of the subject ID, excluded from the table even if numeric}

\item{digits}{how many digits to round to (default = 2)}
}
\value{
a tbl of correlations, means and sds
}
\description{
Generates a table of the correlations and means of numeric columns in a data frame. If data was generated by \code{sim_design} and has a "design" attribute, between, within, dv and id are retrieved from that, unless overridden (use between = 0 to
}
\examples{
get_params(iris, "Species")
}
