% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcols.R
\name{getcols}
\alias{getcols}
\title{Get data columns}
\usage{
getcols(data, ..., as_index = FALSE)
}
\arguments{
\item{data}{the existing tbl}

\item{...}{Columns to get}

\item{as_index}{return the column indices (defaults to name)}
}
\value{
vector of column names or indices
}
\description{
Get columns from a data table by specifying the index, column name as a string, or unquoted column name. Returns the column names or indices.
}
\examples{
getcols(mtcars, 1, cyl, "disp", 5:7)
}
