% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_design}
\alias{interactive_design}
\title{Set design interactively}
\usage{
interactive_design(plot = FALSE)
}
\arguments{
\item{plot}{whether to show a plot of the design}
}
\value{
list
}
\description{
Set design interactively
}
\examples{
des <- interactive_design()
}
