% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_design}
\alias{interactive_design}
\title{Set design interactively}
\usage{
interactive_design(
  output = c("faux", "ANOVApower"),
  plot = faux_options("plot")
)
}
\arguments{
\item{output}{what type of design to output (for faux or ANOVApower)}

\item{plot}{whether to show a plot of the design}
}
\value{
list
}
\description{
Set design interactively
}
\examples{
\dontrun{des <- interactive_design()}
}
