% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_stats.R
\name{long2wide}
\alias{long2wide}
\title{Long to wide format
Converts a table from long to wide format}
\usage{
long2wide(dat, between = c(), within = c(), dv = c(), id = c())
}
\arguments{
\item{dat}{the existing dataframe in long format}

\item{between}{a vector of column names for between-subject factors}

\item{within}{a vector of column names for within-subject factors}

\item{dv}{the column name of the dv}

\item{id}{the column name(s) of the subject ID}
}
\value{
tibble in wide format
}
\description{
Long to wide format
Converts a table from long to wide format
}
\examples{

df_long <- sim_design(2,2, frame_long = TRUE)
df_wide <- long2wide(df_long, "B", "A", "val", "sub_id")

}
