% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{long2wide}
\alias{long2wide}
\title{Long to wide format}
\usage{
long2wide(dat, within = c(), between = c(), dv = c(), id = c())
}
\arguments{
\item{dat}{the long data frame to convert}

\item{within}{the names of the within column(s)}

\item{between}{the names of between column(s) (optional)}

\item{dv}{the name of the DV (value) column}

\item{id}{the names of the column(s) for grouping observations}
}
\value{
the data frame in wide format
}
\description{
Converts data from long format to wide
}
