% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{long2wide}
\alias{long2wide}
\title{Convert data from long to wide format}
\usage{
long2wide(data, within = c(), between = c(), dv = "y", id = "id")
}
\arguments{
\item{data}{the tbl in long format}

\item{within}{the names of the within column(s)}

\item{between}{the names of between column(s) (optional)}

\item{dv}{the name of the DV (value) column}

\item{id}{the names of the column(s) for grouping observations}
}
\value{
a tbl in wide format
}
\description{
Convert data from long to wide format
}
\examples{
df_long <- sim_design(2, 2, long = TRUE)
long2wide(df_long, "A", "B")

}
