% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id.R
\name{make_id}
\alias{make_id}
\title{Make ID}
\usage{
make_id(n = 100, prefix = "S")
}
\arguments{
\item{n}{the number of IDs to generate}

\item{prefix}{the letter prefix to the number}
}
\value{
a vector of IDs
}
\description{
Make IDs with fixed length and a letter prefix for random effects (e.g., S001, S002, ..., S100).
}
\examples{

make_id(20, "SUBJECT_")

}
