% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{OR}
\alias{OR}
\alias{\%||\%}
\title{Piped OR}
\usage{
l \%||\% r
}
\arguments{
\item{l}{LHS.}

\item{r}{RHS.}
}
\value{
LHS if not \code{NULL}, otherwise RHS.
}
\description{
LHS if not \code{NULL}, otherwise RHS
}
\examples{
x <- list(b = 2, c = 3)
x$a \%||\% x$b \%||\% x$c
x$a \%||\% "default_value"

}
