% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot design}
\usage{
plot_design(design, id = "sub_id", dv = "val")
}
\arguments{
\item{design}{A list of design parameters created by check_design() or a data tbl (in long format)}

\item{id}{the column name(s) that identify a unit of analysis}

\item{dv}{the column name that identifies the DV}
}
\value{
plot
}
\description{
\code{plot_design()} plots the specified within and between design
}
\examples{

within <- list(time = c("day", "night"))
between <- list(pet = c("dog", "cat"))
check_design(within, between) \%>\% plot_design()

}
