% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot design}
\usage{
plot_design(input, ...)
}
\arguments{
\item{input}{A list of design parameters created by check_design() or a data tbl (in long format)}

\item{...}{A list of factor names to determine visualisation (see vignette)}
}
\value{
plot
}
\description{
\code{plot_design()} plots the specified within and between design
}
\examples{

within <- list(time = c("day", "night"))
between <- list(pet = c("dog", "cat"))
des <- check_design(within, between, plot = FALSE)
plot_design(des)

data <- sim_design(within, between, plot = FALSE)
plot_design(data)

}
