% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\alias{plot.design}
\alias{plot.faux}
\title{Plot design}
\usage{
plot_design(x, ..., geoms = NULL, palette = "Dark2", labeller = "label_value")

\method{plot}{design}(x, ...)

\method{plot}{faux}(x, ...)
}
\arguments{
\item{x}{A list of design parameters created by check_design() or a data tbl (in long format)}

\item{...}{A list of factor names to determine visualisation (see vignette) in the order color, x, facet row(s), facet col(s)}

\item{geoms}{A list of ggplot2 geoms to display, defaults to "pointrangeSD" (mean ± 1SD) for designs and c("violin", "box") for data, options are: pointrangeSD, pointrangeSE, violin, box, jitter}

\item{palette}{A brewer palette, defaults to "Dark2" (see ggplot2::scale_colour_brewer)}

\item{labeller}{How to label the facets (see ggplot2::facet_grid). "label_value" is used by default.}
}
\value{
plot
}
\description{
Plots the specified within and between design. See \href{../doc/plots.html}{\code{vignette("plots", package = "faux")}} for examples and details.
}
\section{Functions}{
\itemize{
\item \code{plot(design)}: Plotting from a faux design list

\item \code{plot(faux)}: Plotting from a faux data table

}}
\examples{

within <- list(time = c("day", "night"))
between <- list(pet = c("dog", "cat"))
des <- check_design(within, between, plot = FALSE)
plot_design(des)

data <- sim_design(within, between, plot = FALSE)
plot_design(data)
}
