% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_list.R
\name{print.nested_list}
\alias{print.nested_list}
\title{Print Nested List}
\usage{
\method{print}{nested_list}(x, ...)
}
\arguments{
\item{x}{The nested_list string}

\item{...}{Additional parameters for print}
}
\value{
Prints x and returns it invisibly
}
\description{
Print Nested List
}
\keyword{internal}
