% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{readline_check}
\alias{readline_check}
\title{Check readline input}
\usage{
readline_check(prompt, type = c("numeric", "integer", "length",
  "minlength", "maxlength", "exact", "grep"), compare = NULL, ...)
}
\arguments{
\item{prompt}{the prompt for readline}

\item{type}{what type of check to perform, one of c("numeric", "character", "length", "minlength", "maxlength", "exact", "grep")}

\item{compare}{the comparator for exact and (min|max)length types}

\item{...}{other arguments to pass to grep}
}
\value{
the validated result of readline
}
\description{
Check readline input
}
\examples{
readline_check("Type a number: ", "numeric")
readline_check("Type two characters: ", "length", 2)
readline_check("Type at least 3 characters: ", "minlength", 3)
readline_check("Type no more than 4 characters: ", "maxlength", 4)
readline_check("Type a letter and a number: ", "grep", pattern = "^[a-zA-Z]\\\\d$")

}
