% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from design (internal)}
\usage{
sim_data(
  design,
  empirical = FALSE,
  long = faux_options("long"),
  rep = 1,
  nested = TRUE,
  seed = NULL
)
}
\arguments{
\item{design}{A list of design parameters created by check_design()}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}

\item{rep}{the number of data frames to return (default 1); if greater than 1, the returned data frame is nested by rep}

\item{nested}{Whether to nest data frames by rep if rep > 1}

\item{seed}{DEPRECATED use set.seed() instead before running this function}
}
\value{
a tbl
}
\description{
Simulate data from design (internal)
}
\keyword{internal}
