% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design}
\alias{sim_design}
\title{Simulate Data from Design}
\usage{
sim_design(within = list(), between = list(), n = 100, mu = 0,
  sd = 1, r = 0, empirical = FALSE, frame_long = FALSE)
}
\arguments{
\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{mu}{a vector giving the means of the variables (numeric vector of length 1 or vars)}

\item{sd}{the standard deviations of the variables (numeric vector of length 1 or vars)}

\item{r}{the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{frame_long}{Whether the returned dataframe is in wide (default = FALSE) or long (TRUE) format}
}
\value{
dataframe
}
\description{
\code{sim_design} generates a dataframe with a specified within and between design
}
