% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design}
\alias{sim_design}
\title{Generate Data from Design (NOT DONE!!!)}
\usage{
sim_design(within = list(), between = list(), n = 100, cors = 0,
  mu = 0, sd = 1, empirical = FALSE)
}
\arguments{
\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{cors}{the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)}

\item{mu}{a vector giving the means of the variables (numeric vector of length 1 or vars)}

\item{sd}{the standard deviations of the variables (numeric vector of length 1 or vars)}

\item{empirical}{logical. If true, mu, sd and cors specify the empirical not population mean, sd and covariance}
}
\value{
dataframe
}
\description{
\code{sim_design} generates a dataframe with a specified within and between design
}
