% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design}
\alias{sim_design}
\title{Simulate data from design}
\usage{
sim_design(within = list(), between = list(), n = 100, mu = 0,
  sd = 1, r = 0, empirical = FALSE, long = FALSE, dv = list(y =
  "value"), id = list(id = "id"), plot = faux_options("plot"),
  seed = NULL, interactive = FALSE, design = NULL)
}
\arguments{
\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{mu}{the means of the variables}

\item{sd}{the standard deviations of the variables}

\item{r}{the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}

\item{dv}{the name of the dv for long plots (defaults to y)}

\item{id}{the name of the id column (defaults to id)}

\item{plot}{whether to show a plot of the design}

\item{seed}{a single value, interpreted as an integer, or NULL (see set.seed)}

\item{interactive}{whether to run the function interactively}

\item{design}{a design list including within, between, n, mu, sd, r, dv, id}
}
\value{
a tbl
}
\description{
\code{sim_design()} generates a data table with a specified within and between design.
}
