% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design_}
\alias{sim_design_}
\title{Simulate data from design (internal)}
\usage{
sim_design_(design, empirical = FALSE, long = FALSE, seed = NULL)
}
\arguments{
\item{design}{A list of design parameters created by check_design()}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}

\item{seed}{a single value, interpreted as an integer, or NULL (see set.seed)}
}
\value{
a tbl
}
\description{
Simulate data from design (internal)
}
\keyword{internal}
