% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design.R
\name{sim_design_}
\alias{sim_design_}
\title{Simulate Data from Design}
\usage{
sim_design_(design, empirical = FALSE, frame_long = FALSE)
}
\arguments{
\item{design}{A list of design parameters created by check_design()}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{frame_long}{Whether the returned dataframe is in wide (default = FALSE) or long (TRUE) format}
}
\value{
dataframe
}
\description{
\code{sim_design_} generates a dataframe with a specified design
}
\keyword{internal}
