% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{Simulate an existing dataframe}
\usage{
sim_df(data, n = 100, within = c(), between = c(), id = "sub_id",
  dv = "val", empirical = FALSE, long = FALSE, seed = NULL,
  grp_by = NULL)
}
\arguments{
\item{data}{the existing tbl (must be in wide format)}

\item{n}{the number of samples to return per group}

\item{within}{a list of the within-subject columns}

\item{between}{a list of the between-subject columns}

\item{id}{the names of the column(s) for grouping observations}

\item{dv}{the name of the DV (value) column}

\item{empirical}{logical. Passed on to rnorm_multi}

\item{long}{whether to return the data table in long format}

\item{seed}{a single value, interpreted as an integer, or NULL (see set.seed)}

\item{grp_by}{(deprecated; use between)}
}
\value{
a tbl
}
\description{
\code{sim_df} Produces a data table with the same distributions and correlations 
as an existing data table Only returns numeric columns and simulates all numeric 
variables from a continuous normal distribution (for now).
}
\examples{
iris100 <- sim_df(iris, 100)
iris_species <- sim_df(iris, 100, between = "Species")
}
