% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{Simulate an existing dataframe}
\usage{
sim_df(dat, n = 100, between = c(), within = c(),
  empirical = FALSE)
}
\arguments{
\item{dat}{the existing dataframe}

\item{n}{the number of samples to return per group}

\item{between}{a list of the between-subject factors}

\item{within}{a list of the within-subject factors}

\item{empirical}{logical. Passed on to rnorm_multi}
}
\value{
tibble
}
\description{
\code{sim_df} Produces a dataframe with the same distributions and correlations as an existing dataframe. Only returns numeric columns and simulates all numeric variables from a continuous normal distribution (for now).
}
\examples{
iris100 <- sim_df(iris, 100)
iris_species <- sim_df(iris, 100, "Species")
}
