% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{Simulate an existing dataframe}
\usage{
sim_df(
  data,
  n = 100,
  within = c(),
  between = c(),
  id = "id",
  dv = "value",
  empirical = FALSE,
  long = faux_options("long"),
  seed = NULL,
  missing = FALSE,
  sep = faux_options("sep")
)
}
\arguments{
\item{data}{the existing tbl}

\item{n}{the number of samples to return per group}

\item{within}{a list of the within-subject factor columns (if long format)}

\item{between}{a list of the between-subject factor columns}

\item{id}{the names of the column(s) for grouping observations}

\item{dv}{the name of the DV (value) column}

\item{empirical}{Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)}

\item{long}{whether to return the data table in long format}

\item{seed}{DEPRECATED use set.seed() instead before running this function}

\item{missing}{simulate missing data?}

\item{sep}{separator for factor levels}
}
\value{
a tbl
}
\description{
Produces a data table with the same distributions and correlations 
as an existing data table Only returns numeric columns and simulates all numeric variables from a continuous normal distribution (for now).
}
\details{
See \code{vignette("sim_df", package = "faux")} for details.
}
\examples{
iris100 <- sim_df(iris, 100)
iris_species <- sim_df(iris, 100, between = "Species")

# set the names of within factors and (the separator character) 
# if you want to return a long version
longdf <- sim_df(iris, 
                 between = "Species", 
                 within = c("type", "dim"),
                 sep = ".",
                 long = TRUE)
                 
# or if you are simulating data from a table in long format
widedf <- sim_df(longdf, 
                 between = "Species", 
                 within = c("type", "dim"),
                 sep = ".")
}
