% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mixed_df.R
\name{sim_mixed_df}
\alias{sim_mixed_df}
\title{Generate a sample with random intercepts for subjects and items}
\usage{
sim_mixed_df(.data, sub_n = 100, item_n = 25, dv = 1, sub_id = 2,
  item_id = 3)
}
\arguments{
\item{.data}{the existing tbl}

\item{sub_n}{the number of subjects to simulate}

\item{item_n}{the number of items to simulate}

\item{dv}{the column name or index containing the DV}

\item{sub_id}{the column name or index for the subject IDs}

\item{item_id}{the column name or index for the item IDs}
}
\value{
a tbl
}
\description{
\code{sim_mixed_df()} produces a data table with the same distributions of by-subject and by-item random intercepts as an existing data table.
}
\examples{
\donttest{sim_mixed_df(faceratings, 10, 10, "rating", "rater_id", "face_id")}
}
