% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{wide2long}
\alias{wide2long}
\title{Wide to long format}
\usage{
wide2long(.data, within_factors = c(), within_cols = c(), dv = "val",
  id = NULL, sep = "[^[:alnum:]]+")
}
\arguments{
\item{.data}{the tbl in wide format}

\item{within_factors}{the names of the within factors}

\item{within_cols}{the names (or indices) of the within-subject (value) columns}

\item{dv}{the name of the dv column (defaults to "val")}

\item{sep}{Separator for within-columns (see tidyr::separate)}
}
\value{
a tbl in long format
}
\description{
Converts data from wide format to long
}
\examples{
wide2long(iris, c("Feature", "Measure"), 1:4)

}
